# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sabi Health Backend API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: Canonical frontend and operations reference generated from the registered Express routes and validation contracts. All currency values ending in Minor are NGN minor units (kobo).
servers:
  - url: /
    description: Current environment
...
paths:
...
  /api/v1/medications:
    get:
      tags:
        - Medications
      summary: List medications
      operationId: get__api_v1_medications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Medications
      summary: Create medication
      operationId: post__api_v1_medications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Medication"
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medications/{id}:
    get:
      tags:
        - Medications
      summary: Get medication
      operationId: get__api_v1_medications_id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags:
        - Medications
      summary: Update medication
      operationId: put__api_v1_medications_id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags:
        - Medications
      summary: Delete medication
      operationId: delete__api_v1_medications_id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/v1/medication-schedules:
    get:
      tags:
        - Medication Schedules
      summary: My medicines and their schedules
      description: Live schedules, plus ones stopped or completed in the last 30 days.
      operationId: get__api_v1_medication_schedules
      security:
        - bearerAuth: []
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Medication Schedules
      summary: Set up a medicine
      description: From an issued prescription item (`prescriptionItemId`; name and dose come from the prescription) or a medicine the patient adds (`name`). "As needed" medicines are listed without times or reminders. Doses are created up to 48 hours ahead, each with a reminder.
      operationId: post__api_v1_medication_schedules
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                prescriptionItemId:
                  type: string
                  format: uuid
                name:
                  type: string
                  maxLength: 120
                dosage:
                  type: string
                  maxLength: 64
                instructions:
                  type: string
                  maxLength: 500
                asNeeded:
                  type: boolean
                times: &a3
                  type: array
                  maxItems: 8
                  items:
                    type: string
                    pattern: ^([01]\d|2[0-3]):[0-5]\d$
                  description: Clock times in the patient time zone, e.g. ["08:00","20:00"].
                startDate: &a2
                  type: string
                  format: date
                endDate: *a2
                remindersEnabled:
                  type: boolean
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medication-schedules/suggestions:
    get:
      tags:
        - Medication Schedules
      summary: Prescribed medicines not set up yet
      description: Issued prescription items with times suggested from the prescribed frequency and a last day from the duration. "As needed" medicines have no suggested times.
      operationId: get__api_v1_medication_schedules_suggestions
      security:
        - bearerAuth: []
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medication-schedules/{id}:
    patch:
      tags:
        - Medication Schedules
      summary: Change a medicine
      description: "Any change creates a new version: future doses and reminders of the old version are cancelled and new ones made. Pausing stops reminders until resumed."
      operationId: patch__api_v1_medication_schedules_id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                dosage:
                  type: string
                instructions:
                  type: string
                times: *a3
                endDate: *a2
                remindersEnabled:
                  type: boolean
                paused:
                  type: boolean
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medication-schedules/{id}/stop:
    post:
      tags:
        - Medication Schedules
      summary: Stop a medicine
      description: Cancels future doses and reminders; past doses stay on record. Cancelling a prescription stops its medicines automatically.
      operationId: post__api_v1_medication_schedules_id_stop
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200": *a1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medication-schedules/doses:
    get:
      tags:
        - Medication Schedules
      summary: Doses for a day
      description: Doses on one calendar day in the patient time zone (today by default), each with what the patient reported (NOT_CONFIRMED, TAKEN, SKIPPED) and the reminder state. A reminder being delivered or read never counts as taken.
      operationId: get__api_v1_medication_schedules_doses
      security:
        - bearerAuth: []
      parameters:
        - name: day
          in: query
          schema: *a2
      responses:
        "200": *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medication-schedules/doses/{id}/taken:
    post:
      tags:
        - Medication Schedules
      summary: Record a dose as taken
      description: "Idempotent: recording it again returns `alreadyRecorded: true`. Cancels the dose's pending reminder. Allowed from two hours before the dose is due."
      operationId: post__api_v1_medication_schedules_doses_id_taken
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200": *a1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/medication-schedules/doses/{id}/skipped:
    post:
      tags:
        - Medication Schedules
      summary: Record a dose as skipped
      description: Same rules as taken.
      operationId: post__api_v1_medication_schedules_doses_id_skipped
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200": *a1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Use an access token. Example: Bearer <access-token>"
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        status:
          const: error
        message:
          type: string
        code:
          type: string
      required:
        - status
        - message
    Success:
      type: object
      additionalProperties: false
      properties:
        status:
          const: success
        data: {}
      required:
        - status
...
    Medication:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        instructions:
          type: string
        time:
          type: string
      required:
        - name
        - instructions
        - time
    MedicationUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        instructions:
          type: string
        time:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource is absent or hidden by authorization scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: State conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Safe module-scoped server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...